requirements.txt
Flask==3.0.3
gunicorn==22.0.0
moviepy==1.0.3
edge-tts
curl_cffi
beautifulsoup4
Pillow
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests
